Various
- Renamed and re-organized league code
- Added support for new env info
