+++ src/envs/__init__.py
@@ -1,7 +1,7 @@
 import logging
 from functools import partial
 
-from smac.env import MultiAgentEnv, StarCraft2Env
+from smac.env import StarCraft2Env
 
 from multiagent.environment import MAEnv
 from multiagent.scenarios import team
@@ -10,14 +10,13 @@ from bin.team_plans_example import TWO_TEAMS_SIZE_TWO_SYMMETRIC_HETEROGENEOUS
 import sys
 import os
 
-
-def smac_env(env, **kwargs) -> MultiAgentEnv:
-    return env(**kwargs)
-
-
 if sys.platform == "linux":
     os.environ.setdefault("SC2PATH",
                           os.path.join(os.getcwd(), "3rdparty", "StarCraftII"))
+
+
+def smac_env(env, **kwargs) -> StarCraft2Env:
+    return env(**kwargs)
 
 
 def ma_env(env, **kwargs) -> MAEnv:
@@ -28,12 +27,12 @@ def ma_env(env, **kwargs) -> MAEnv:
     world = scenario.make_teams_world(grid_size=10.0)
     # create multi-agent environment
     return env(world=world,
-                 reset_callback=scenario.reset_world,
-                 reward_callback=scenario.reward,
-                 observation_callback=scenario.observation,
-                 info_callback=None,
-                 done_callback=scenario.done,
-                 log=True, log_level=logging.ERROR)
+               reset_callback=scenario.reset_world,
+               reward_callback=scenario.reward,
+               observation_callback=scenario.observation,
+               info_callback=None,
+               done_callback=scenario.done,
+               log=True, log_level=logging.ERROR)
 
 
 REGISTRY = {}
