+++ src/league/roles/exploiters.py
@@ -1,0 +1,80 @@
+class MainExploiter(Player):
+
+    def __init__(self, race, agent, payoff):
+        self.agent = Agent(race, agent.get_weights())
+        self._initial_weights = agent.get_weights()
+        self._payoff = payoff
+        self._race = agent.race
+        self._checkpoint_step = 0
+
+    def get_match(self):
+        main_agents = [
+            player for player in self._payoff.players
+            if isinstance(player, MainPlayer)
+        ]
+        opponent = np.random.choice(main_agents)
+
+        if self._payoff[self, opponent] > 0.1:
+            return opponent, True
+
+        historical = [
+            player for player in self._payoff.players
+            if isinstance(player, Historical) and player.parent == opponent
+        ]
+        win_rates = self._payoff[self, historical]
+
+        return np.random.choice(
+            historical, p=prioritized_fictitious_self_play(win_rates, weighting="variance")), True
+
+    def checkpoint(self):
+        self.agent.set_weights(self._initial_weights)
+        self._checkpoint_step = self.agent.get_steps()
+        return self._create_checkpoint()
+
+    def ready_to_checkpoint(self):
+        steps_passed = self.agent.get_steps() - self._checkpoint_step
+        if steps_passed < 2e9:
+            return False
+
+        main_agents = [
+            player for player in self._payoff.players
+            if isinstance(player, MainPlayer)
+        ]
+        win_rates = self._payoff[self, main_agents]
+        return win_rates.min() > 0.7 or steps_passed > 4e9
+
+
+class LeagueExploiter(Player):
+
+    def __init__(self, race, agent, payoff):
+        self.agent = Agent(race, agent.get_weights())
+        self._initial_weights = agent.get_weights()
+        self._payoff = payoff
+        self._race = agent.race
+        self._checkpoint_step = 0
+
+    def get_match(self):
+        historical = [
+            player for player in self._payoff.players
+            if isinstance(player, Historical)
+        ]
+        win_rates = self._payoff[self, historical]
+        return np.random.choice(
+            historical, p=prioritized_fictitious_self_play(win_rates, weighting="linear_capped")), True
+
+    def checkpoint(self):
+        if np.random.random() < 0.25:
+            self.agent.set_weights(self._initial_weights)
+        self._checkpoint_step = self.agent.get_steps()
+        return self._create_checkpoint()
+
+    def ready_to_checkpoint(self):
+        steps_passed = self._agent.get_steps() - self._checkpoint_step
+        if steps_passed < 2e9:
+            return False
+        historical = [
+            player for player in self._payoff.players
+            if isinstance(player, Historical)
+        ]
+        win_rates = self._payoff[self, historical]
+        return win_rates.min() > 0.7 or steps_passed > 4e9
